Consolidate auth monitoring metrics
